Cleaned GraphUnzip from remnants of the old GraphUnzip-HiC
